Keep OAuth regression checks CI-safe
